the authentication updates

+++ mss/auth_sys/delete.php
@@ -6,9 +6,10 @@ $user_id = $_SESSION['email'];
 database_connect();
 $result = retrieve_user($user_id);
 database_disconnect();
-if ($result[0]['admin'] == 1)
+
+$email = $_REQUEST['email'];
+if (($result[0]['admin'] == 1) && ($email != $user_id))
 {
-   $email = $_GET['email'];
    database_connect();
    $result = delete_user($email);
    database_disconnect();
